pyproject.toml
[project]
name = "solo-founder-agents"
version = "0.1.0"
description = "Multi-agent AI system for solo founders to build digital products"
authors = [
    {name = "vovkins",email = "vovkins@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "crewai (>=1.12.2,<2.0.0)",
    "pygithub (>=2.9.0,<3.0.0)",
    "python-telegram-bot (>=22.7,<23.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.8,<0.16.0)"
]
